# Swagger API Documentaion.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a great resource for using endpoints of the monitorMyRehab API. We use a general swagger UI for all endpoints. Completed routes doc - users, admins, appointment, healthprofessionalTypes, sessions",
    "version": "1.0.0",
    "title": "MonitorMyRehab API",
    "termsOfService": "https://monitormyrehab.com/terms/",
    "contact": {
      "email": "info@monitormyrehab.com"
    },
    "license": {
      "name": "UNLICENSED",
      "url": "https://monitormyrehab.com/privacy/"
    }
  },
  "host": "api.mmr-meet.click",
  "basePath": "/api/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/users/login/": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Login with user's details",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginDetails"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/LoginResponse"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/users/register/": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Register an user",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RegisterDetails"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/users/": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get all users",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an admin authorization token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Add user",
        "description": "Add user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an admin authorization token",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Add user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/users/paginate": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get paginated users",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an admin authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/users/count": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get all users count",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an admin authorization token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get one user by id",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an admin authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "UID of the user to get",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update user by id",
        "description": "Update user details",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an admin authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "UID of the user to get",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete user with id",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an admin authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "UID of the user to get",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/users/paginate/": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get all users in pagination mode",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an admin authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/users/users/": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get all healthprofessional's users",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an healthprofessional authorization token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Add healthprofessional user",
        "description": "Add healthprofessional user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an healthprofessional authorization token",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Add user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/users/users/paginate": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get paginated users",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an health professional authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/users/users/count": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get all healthprofessional's users count",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an healthprofessional authorization token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/users/users/paginate/": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get all healthprofessional's users in pagination mode",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an healthprofessional authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/users/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get health professional's user with id",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an healthprofessional authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "UID of the user to get",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update user by id",
        "description": "Update user details",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an healthprofessional authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "UID of the user to get",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete health professional's user with id",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an healthprofessional authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "UID of the user to get",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
...
    "/appointments/users": {
      "get": {
        "tags": [
          "appointments"
        ],
        "summary": "Get all user appointments - health professional access.",
        "description": "This endpoint returns all appointments in the system. We encourage you to use the pagination endpoint for appointments.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "a user authorization token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PostSuccessArray"
            }
          },
          "403": {
            "description": "Access Forbiden",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/AppointmentsError"
            }
          }
        }
      },
      "post": {
        "tags": [
          "appointments"
        ],
        "summary": "Create a user appointment",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "a user authorization token",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AppointmentCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/appointments/users/paginate": {
      "get": {
        "tags": [
          "appointments"
        ],
        "summary": "Get paginated user appointments - admin access.",
        "description": "This endpoint returns all user appointments in the system. We encourage you to use the pagination endpoint for appointments.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an user authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PostSuccessArray"
            }
          },
          "403": {
            "description": "Access Forbiden",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/AppointmentsError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "LoginResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "description": "Shows status of request"
        },
        "token": {
          "type": "string",
          "description": ""
        },
        "route": {
          "type": "string",
          "description": ""
        }
      }
    },
    "AccessError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "description": "Shows status of request"
        },
        "message": {
          "type": "string",
          "description": "",
          "enum": [
            "Unauthorized Access"
          ]
        },
        "route": {
          "type": "string",
          "description": ""
        }
      }
    },
    "ServerError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "description": "Shows status of request"
        },
        "message": {
          "type": "string",
          "description": "",
          "enum": [
            "Server Error"
          ]
        },
        "route": {
          "type": "string",
          "description": ""
        }
      }
    },
...
    "LoginDetails": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "description": "User email"
        },
        "password": {
          "type": "string",
          "description": "User password"
        },
        "remember": {
          "type": "boolean",
          "description": "Remember token for longer or not."
        }
      }
    },
...
  }
}